+++ pytsite/core/odm/__init__.py
@@ -4,12 +4,9 @@ __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-import pymongo as _pymongo
-from . import _model, _field, _manager, _finder, _validation, _error
-
-
-
 def _app_update_event():
+    """'app.update' event handler.
+    """
     from pytsite.core import console
     console.run_command('odm:reindex')
 
@@ -19,24 +16,15 @@ def __init():
     from . import _console_commands
 
     lang.register_package(__name__)
-
     console.register_command(_console_commands.RebuildIndices())
-
     events.listen('app.update', _app_update_event)
 
 __init()
 
+
 # Public API
+from pymongo import ASCENDING as I_ASC, DESCENDING as I_DESC, GEO2D as I_GEO2D
+from . import _field as field, _validation as validation, _error as error
 from ._model import Model
-I_ASC = _pymongo.ASCENDING
-I_DESC = _pymongo.DESCENDING
-I_GEO2D = _pymongo.GEO2D
-Finder = _finder.Finder
-field = _field
-validation = _validation
-error = _error
-register_model = _manager.register_model
-find = _manager.find
-dispense = _manager.dispense
-get_by_ref = _manager.get_by_ref
-resolve_ref = _manager.resolve_ref
+from ._finder import Finder
+from ._functions import register_model, find, dispense, get_by_ref, resolve_ref