+++ pytsite/core/widget/_input.py
@@ -166,4 +166,4 @@ class CKEditor(_base.Base):
     def render(self) -> str:
         """Render the widget.
         """
-        return self._group_wrap(_html.TextArea(name=self._uid))
+        return self._group_wrap(_html.TextArea(self.get_value(), name=self._uid))